% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BmusClus.R
\name{BmusClus}
\alias{BmusClus}
\title{Cluster assignment for the experimental data}
\usage{
BmusClus(Bmus, Cluster)
}
\arguments{
\item{Bmus}{Best Matching Unit assignment to the experimental data}

\item{Cluster}{Vector containing cluster number assignment for prototypes}
}
\value{
A vector containing the cluster assignment to experimental data
}
\description{
Generate a vector containing the cluster assignment to experimental data
}
\author{
S. Licen
}
