% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxClus.R
\name{BoxClus}
\alias{BoxClus}
\title{Boxplot of prototype variables split by cluster and variable}
\usage{
BoxClus(Dms, codebook, Cluster)
}
\arguments{
\item{Dms}{A vector of length 2, where the first argument specifies the number of rows and the second the number of columns of plots (see mfrow in par)}

\item{codebook}{Prototype codebook de-normalized}

\item{Cluster}{Vector containing cluster number assignment for prototypes}
}
\value{
Boxplot of prototype variables split by cluster
}
\description{
Boxplot function is used, box whiskers are omitted
}
\author{
S. Licen
}
