% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxUnits.R
\name{BoxUnits}
\alias{BoxUnits}
\title{Boxplot of prototype variables split by cluster}
\usage{
BoxUnits(codebook, Cluster, Ylim = NA, pitch = NA, xdim = 0.75)
}
\arguments{
\item{codebook}{Prototype codebook normalized by variable}

\item{Cluster}{Vector containing cluster number assignment for prototypes}

\item{Ylim}{Vector of length 2 for y-axis limits}

\item{pitch}{Vector containing the position of horizontal grid lines}

\item{xdim}{x axes label dimensions}
}
\value{
Boxplot of prototype variables split by cluster
}
\description{
Boxplot function is used, box whiskers are omitted
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\seealso{
boxplot
}
\author{
Sabina Licen
}
