% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusCol.R
\name{ClusCol}
\alias{ClusCol}
\title{Custom color sequence for clusters}
\usage{
ClusCol(Centroids, Cluster, colSeq = rainbow(nrow(data.frame(Centroids))))
}
\arguments{
\item{Centroids}{Centroids matrix}

\item{Cluster}{Vector containing cluster number assignment for prototypes}

\item{colSeq}{Color sequence for the clusters}
}
\value{
A vector of colors with length equal to Cluster
}
\description{
Generate the sequence of colors to plot the SOM map according clusters
}
\author{
S. Licen
}
