% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_indexR.R
\name{db_indexR}
\alias{db_indexR}
\title{Evaluate Davis-Bouldin index}
\usage{
db_indexR(codebook, k_best, c_best)
}
\arguments{
\item{codebook}{SOM codebook}

\item{k_best}{Vector with cluster number assignment for each sample}

\item{c_best}{Matrix with cluster centroids}
}
\value{
The mean DB-index for the clustering
}
\description{
The function has been coded in R code starting from db_index.m
script present in somtoolbox for Matlab by Vesanto
}
\author{
S. Licen
}
