% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq.R
\name{Freq}
\alias{Freq}
\title{Percentage frequency for each cluster}
\usage{
Freq(Cluster)
}
\arguments{
\item{Cluster}{Vector containing cluster number assignment for experimental data}
}
\value{
A data frame containing the percentage frequency of each cluster
}
\description{
Percentage frequency for each cluster
}
\author{
S. Licen
}
