% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqM.R
\name{FreqM}
\alias{FreqM}
\title{Monthly percentage frequency for each cluster}
\usage{
FreqM(Date, Cluster)
}
\arguments{
\item{Date}{Vector containing date/time variable for experimental data}

\item{Cluster}{Vector containing cluster number assignment for experimental data}
}
\value{
A data frame containing the monthly percentage frequency of each cluster
}
\description{
Monthly percentage frequency for each cluster
}
\author{
S. Licen
}
