% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hexa.R
\name{Hexa}
\alias{Hexa}
\title{Function to draw an hexagon around a point}
\usage{
Hexa(x, y, color = NA, border = "gray", unitcell = 1)
}
\arguments{
\item{x}{X-coordinate of the hexagon center}

\item{y}{Y-coordinate of the hexagon center}

\item{color}{Filling color of the hexagon (default NA)}

\item{border}{Border color of the hexagon (default "gray")}

\item{unitcell}{The distance side to side between two parallel sides of the hexagon (default 1)}
}
\value{
This function draws an hexagon on a plot
}
\description{
Draws an hexagon around a point of x and y coordinates
}
\author{
Sabina Licen
}
