% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hexagons.R
\name{Hexagons}
\alias{Hexagons}
\title{Function to draw an hexagonal SOM map}
\usage{
Hexagons(Coords, Row, Col, color = NA, border = "gray", unitcell = 1)
}
\arguments{
\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}

\item{color}{Filling color of the hexagons (default NA)}

\item{border}{Border color of the hexagons (default "gray")}

\item{unitcell}{The distance side to side between two parallel sides of the hexagon (default 1)}

\item{x}{X-coordinates of the hexagon centers}

\item{y}{Y-coordinates of the hexagon centers}
}
\value{
A hexagonal SOM map
}
\description{
Draw an hexagon around a point of x and y coordinates
}
\examples{
Coord<-CodeCoord(10,5)
Hexagons(Coord$XC,Coord$YC,10,5)
}
\author{
S. Licen
}
