% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexagonsVar.R
\name{HexagonsVar}
\alias{HexagonsVar}
\title{Heatmaps}
\usage{
HexagonsVar(Dms, codebook, Coords, Row, Col)
}
\arguments{
\item{Dms}{A vector of length 2, where the first argument specifies the number of rows and the second the number of columns of plots (see mfrow in par)}

\item{codebook}{SOM codebook}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}

\item{Coord}{Prototype coordinates for plotting the map}
}
\value{
Plot a SOM map for the values of each modeled variable using a grayscale according to quartiles
}
\description{
Multiple plots that show the distribution of the modeled variables on the map
}
\details{
The function plots a SOM map for the values of each modeled variable using a grayscale according to quartiles,
from white (lower outliers), followed by grayscale (quartiles) and black (upper outiliers).
The outilers and quartiles are evaluated by boxplot function applying default parameters.
}
\author{
S. Licen
}
