% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexaHits.R
\name{HexaHits}
\alias{HexaHits}
\title{Hits distribution on the SOM map}
\usage{
HexaHits(hits, Coord, Row, Col, color = "black")
}
\arguments{
\item{hits}{Vector with number of hits for each prototype}

\item{Coord}{Prototype coordinates for plotting the map}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}

\item{color}{color filling of the hexagons}
}
\value{
Plot a SOM map with filled hexagons according to the number of hits
}
\description{
Plot a SOM map with filled hexagons according to the number of hits
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
