% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexaQerrs.R
\name{HexaQerrs}
\alias{HexaQerrs}
\title{Realtive quantization error distribution on the SOM map}
\usage{
HexaQerrs(bmus, qerrs, Coord, Row, Col, col = "black")
}
\arguments{
\item{bmus}{Vector with Best Matching Unit for each experimental sample}

\item{qerrs}{Vector with quantization error for each experimental sample}

\item{Coord}{Prototype coordinates for plotting the map}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
Plot a SOM map with filled hexagons according to the realtive quantization error
}
\description{
Plot a SOM map with realtive quantization error plotted as grayscale according to quartiles
}
\details{
The function evaluate the relative quantization error for each prototype
dividing the sum of quantization errors for experimental samples represented by the single
prototype by the number of hits of the same prototype,
then plots a SOM map with with filled hexagons according to the realtive quantization error
}
\author{
S. Licen
}
