% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexaQerrsQuant.R
\name{HexaQerrsQuant}
\alias{HexaQerrsQuant}
\title{Realtive quantization error distribution on the SOM map}
\usage{
HexaQerrsQuant(bmus, qerrs, Coord, Row, Col)
}
\arguments{
\item{bmus}{Vector with Best Matching Unit for each experimental sample}

\item{qerrs}{Vector with quantization error for each experimental sample}

\item{Coord}{Prototype coordinates for plotting the map}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
Plot a SOM map with realtive quantization error represented as grayscale according to quartiles
}
\description{
Plot a SOM map with realtive quantization error plotted as grayscale according to quartiles
}
\details{
The function evaluate the relative quantization error for each prototype
dividing the sum of quantization errors for experimental sample represented by the single
prototype by the number of hits of the same prototype,
then plots a SOM map with the realtive quantization error represented as grayscale according to quartiles,
from white (lower outliers) followed by grayscale (quartiles) and black (upper outiliers).
The outilers and quartiles are evaluated by boxplot function applying default parameters.
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\seealso{
boxplot
}
\author{
S. Licen
}
