% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramQuant.R
\name{paramQuant}
\alias{paramQuant}
\title{Basic statistics of values present in the input vector}
\usage{
paramQuant(param)
}
\arguments{
\item{param}{Numeric vector}
}
\value{
A table which contains basic statistics for the input vector
}
\description{
Generate basic statistics for the input vector
}
\details{
The outilers and quartiles are evaluated by boxplot function applying default parameters.
}
\examples{
library(datasets)
paramQuant(iris[,1])
}
\seealso{
boxplot
}
\author{
Sabina Licen
}
