% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_dimR.R
\name{som_dimR}
\alias{som_dimR}
\title{Calculate map dimensions}
\usage{
som_dimR(dataset, type = "regular")
}
\arguments{
\item{dataset}{Experimental data}

\item{type}{Either "regular", "small" or "big" map (default ="regular")}
}
\value{
This function returns a list containing the number of rows, columns and overall map units
}
\description{
Generate SOM map dimensions according to Vesanto heuristic rules
}
\author{
S. Licen
}
