% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_initR.R
\name{som_initR}
\alias{som_initR}
\title{Calculate initialization matrix for SOM training}
\usage{
som_initR(dataset, Row, Col, munits)
}
\arguments{
\item{dataset}{Experimental data}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}

\item{munits}{Number of SOM map units (Row*Col)}
}
\value{
This function returns an initialization matrix for SOM training
}
\description{
Generate SOM map dimensions according to Vesanto heuristic rules related to the first two
eigenvalues of the experimental data and their related eigenvectors
}
\author{
S. Licen
}
