% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_kmeansRProg.R
\name{som_kmeansRProg}
\alias{som_kmeansRProg}
\title{K-means algorithm on codebook for Shiny}
\usage{
som_kmeansRProg(codebook, k, epochs, seed = NULL)
}
\arguments{
\item{codebook}{SOM codebook}

\item{k}{Number of clusters}

\item{epochs}{Number of training epochs}

\item{seed}{Number for set.seed function}
}
\value{
This function returns a list containing the cluster number assignment for each sample,
the cluster centroids and the total quantization error
}
\description{
The training is run for a custom number of epochs for k number of clusters
}
\details{
The function has been coded in R code starting from som_kmeans.m
script present in somtoolbox for Matlab by Vesanto
}
\author{
S. Licen
}
