% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_mdistR.R
\name{som_mdistR}
\alias{som_mdistR}
\title{Evaluate pairwise distance matrix for the given codebook}
\usage{
som_mdistR(codebook)
}
\arguments{
\item{codebook}{SOM codebook}
}
\value{
Distance matrix (used in db_indexR function)
}
\description{
The function has been coded in R code starting from som_mdist.m
script present in somtoolbox for Matlab by Vesanto
}
\author{
S. Licen
}
