% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_umatR.R
\name{som_umatR}
\alias{som_umatR}
\title{Unified distance matrix for the SOM map}
\usage{
som_umatR(codebook, Row, Col)
}
\arguments{
\item{codebook}{SOM codebook}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
The unified distance matrix for the SOM map
}
\description{
The function has been coded in R code starting from som_umat.m
script present in somtoolbox for Matlab by Vesanto and adapted for the use in the shiny app
}
\references{
{J. Vesanto, J. Himberg, E. Alhoniemi, J. Parhankagas, SOM Toolbox for Matlab
5, Report A57, 2000, Available at: www.cis.hut.fi/projects/somtoolbox/package/papers/techrep.pdf;
A. Ultsch, H.P. Siemon, Proceedings of International Neural Network Conference
(INNC?90), Kluwer academic Publishers, Dordrecht, 1990, pp. 305?308.
}
}
\author{
Sabina Licen, Pierluigi Barbieri
}
