% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomEnvGUI.R
\name{SomEnvGUI}
\alias{SomEnvGUI}
\title{The function starts the SOMEnv GUI}
\usage{
SomEnvGUI()
}
\value{
This function starts the graphical user interface with the default system browser.
The main help suggestion for using the tool are embedded in the GUI
}
\description{
The function starts the SOMEnv GUI
}
\examples{
\dontrun{
SomEnvGUI()
}
}
\references{
{Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan
            McPherson (2019). shiny: Web Application Framework for R. R package
            version 1.4.0. https://CRAN.R-project.org/package=shiny
}

seealso shiny
}
\author{
Sabina Licen, Marco Franzon, Tommaso Rodani
}
