% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOMtopol.R
\name{SOMtopol}
\alias{SOMtopol}
\title{Topographical error for the SOM map}
\usage{
SOMtopol(dataset, codebook, grid)
}
\arguments{
\item{dataset}{Experimental data used for training the map}

\item{codebook}{SOM codebook}

\item{grid}{SOM grid expressed as a matrix of x and y coordinates of the map units}
}
\value{
This function returns the topographical error
}
\description{
Calculate topographical error for the SOM map
}
\references{
{Clark, S., Sisson, S.A., Sharma, A. (2020) Adv. Water Resour. 143, art. no. 103676 DOI: 10.1016/j.advwatres.2020.103676}
}
\author{
Sabina Licen
}
