% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UmatGraph.R
\name{UmatGraph}
\alias{UmatGraph}
\title{U-matrix plot}
\usage{
UmatGraph(umat, Row, Col, colorscale = c("bw", "gs"))
}
\arguments{
\item{umat}{Unified Distance Matrix}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
Plot of Unified Distance Matrix using a grayscale according to quartiles
}
\description{
Plot of Unified Distance Matrix using a grayscale according to quartiles
}
\details{
The function plots a SOM map for the values of each modeled variable using a grayscale according to quartiles,
from dark gray (lower distances) to light gray (higher distances).
The quartiles are evaluated by boxplot function applying default parameters.
}
\author{
S. Licen
}
