% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BmusCentr.R
\name{BmusCentr}
\alias{BmusCentr}
\title{BMUs of the cluster centroids}
\usage{
BmusCentr(centroids, som_model, k)
}
\arguments{
\item{centroids}{Centroids array (output of kmeans_clustersR function)}

\item{som_model}{An object of class kohonen}

\item{k}{Number of clusters}
}
\value{
An array containing the BMU for each centroid
}
\description{
The function finds the Best Matching Units of the cluster centroids
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
