% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BmusClus.R
\name{BmusClus}
\alias{BmusClus}
\title{Cluster assignment for the experimental data}
\usage{
BmusClus(Bmus, Cluster)
}
\arguments{
\item{Bmus}{Best Matching Unit assignment to the experimental data}

\item{Cluster}{Vector containing cluster number assignment for prototypes}
}
\value{
A vector containing the cluster assignment to experimental data
}
\description{
Generate a vector containing the cluster assignment to experimental data
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
