% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusCol.R
\name{ClusCol}
\alias{ClusCol}
\title{Custom color sequence for clusters}
\usage{
ClusCol(Centroids, Cluster, colSeq = rainbow(nrow(data.frame(Centroids))))
}
\arguments{
\item{Centroids}{Centroids matrix}

\item{Cluster}{Vector containing cluster number assignment for prototypes}

\item{colSeq}{Color sequence for the clusters}
}
\value{
A vector of colors with length equal to Cluster
}
\description{
Generate the sequence of colors to plot the SOM map according to clusters
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
