% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeCoord.R
\name{CodeCoord}
\alias{CodeCoord}
\title{Prototype coordinates for graph}
\usage{
CodeCoord(Row, Col)
}
\arguments{
\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item X
 \item Y
}
}
\description{
Generate X and Y coordinates for plotting a SOM map shaped according to Vesanto visualization fashion
}
\examples{
Coord<-CodeCoord(10,5)
}
\references{
{J. Vesanto, J. Himberg, E. Alhoniemi, J. Parhankagas, SOM Toolbox for Matlab
5, Report A57, 2000, Available at: www.cis.hut.fi/projects/somtoolbox/package/papers/techrep.pdf;
Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen, Pierluigi Barbieri
}
