% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_indexR.R
\name{db_indexR}
\alias{db_indexR}
\title{Evaluate Davis-Bouldin index for the cluster split of data input}
\usage{
db_indexR(codebook, k_best, c_best)
}
\arguments{
\item{codebook}{SOM codebook}

\item{k_best}{Vector with cluster number assignment for each sample}

\item{c_best}{Matrix with cluster centroids}
}
\value{
The mean DB-index for the clustering
}
\description{
The function has been coded in R code starting from db_index.m
script present in somtoolbox for Matlab by Vesanto and adapted for the use in the shiny app
}
\references{
{J. Vesanto, J. Himberg, E. Alhoniemi, J. Parhankagas, SOM Toolbox for Matlab
5, Report A57, 2000, Available at: www.cis.hut.fi/projects/somtoolbox/package/papers/techrep.pdf
}
}
\seealso{
som_mdistR, kmeans_clustersRProg
}
\author{
Sabina Licen, Pierluigi Barbieri
}
