% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq.R
\name{Freq}
\alias{Freq}
\title{Percentage frequency for each cluster}
\usage{
Freq(Cluster)
}
\arguments{
\item{Cluster}{Vector containing cluster number assignment for experimental data}
}
\value{
A data frame containing the percentage frequency of each cluster
}
\description{
Percentage frequency for each cluster
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
