% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqD.R
\name{FreqD}
\alias{FreqD}
\title{Daily percentage frequency for each cluster}
\usage{
FreqD(Date, Cluster, Total = 1440)
}
\arguments{
\item{Date}{Vector containing date/time variable for experimental data}

\item{Cluster}{Vector containing cluster number assignment for experimental data}

\item{Total}{Number of observations per day}
}
\value{
A data frame containing the daily percentage frequency of each cluster
}
\description{
Daily percentage frequency for each cluster
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
