% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqM.R
\name{FreqM}
\alias{FreqM}
\title{Monthly percentage frequency for each cluster}
\usage{
FreqM(Date, Cluster)
}
\arguments{
\item{Date}{Vector containing date/time variable for experimental data}

\item{Cluster}{Vector containing cluster number assignment for experimental data}
}
\value{
A data frame containing the monthly percentage frequency of each cluster
}
\description{
Monthly percentage frequency for each cluster
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
