% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexagonsClus.R
\name{HexagonsClus}
\alias{HexagonsClus}
\title{SOM map with clusters}
\usage{
HexagonsClus(
  Centroids,
  Cluster,
  BCentr,
  Coord,
  Row,
  Col,
  colSeq = rainbow(nrow(Centroids))
)
}
\arguments{
\item{Centroids}{Centroids matrix}

\item{Cluster}{Vector containing cluster number assignment for prototypes}

\item{BCentr}{Best Matching Unit of the cluster centroids}

\item{Coord}{Prototype coordinates for plotting the map}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}

\item{colSeq}{Color sequence for the clusters}
}
\value{
A SOM map colored according to cluster splitting
}
\description{
Generates a SOM map colored according to cluster splitting
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\author{
Sabina Licen
}
