% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexagonsVar.R
\name{HexagonsVar}
\alias{HexagonsVar}
\title{Heatmaps}
\usage{
HexagonsVar(Dms, codebook, Coords, Row, Col)
}
\arguments{
\item{Dms}{A vector of length 2, where the first argument specifies the number of rows and the second the number of columns of plots (see mfrow in par)}

\item{codebook}{SOM codebook}

\item{Coords}{Prototype coordinates for plotting the map}

\item{Row}{Number of SOM map rows}

\item{Col}{Number of SOM map columns}
}
\value{
SOM map plots for the values of each modeled variable using a grayscale according to quartiles
}
\description{
Multiple plots that show the distribution of the modeled variables on the SOM map
}
\details{
The function plots a SOM map for the values of each modeled variable using a grayscale according to quartiles,
from white (lower outliers), followed by grayscale (quartiles) and black (upper outiliers).
The outilers and quartiles are evaluated by boxplot function applying default parameters.
}
\references{
{Licen, S., Cozzutto, S., Barbieri, P. (2020) Aerosol Air Qual. Res., 20 (4), pp. 800-809. DOI: 10.4209/aaqr.2019.08.0414
}
}
\seealso{
boxplot, par
}
\author{
Sabina Licen
}
