% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_mdistR.R
\name{som_mdistR}
\alias{som_mdistR}
\title{Evaluate pairwise distance matrix for the given codebook}
\usage{
som_mdistR(codebook)
}
\arguments{
\item{codebook}{SOM codebook}
}
\value{
Distance matrix
}
\description{
The function has been coded in R code starting from som_mdist.m
script present in somtoolbox for Matlab by Vesanto and adapted for the use in the shiny app
}
\references{
{J. Vesanto, J. Himberg, E. Alhoniemi, J. Parhankagas, SOM Toolbox for Matlab
5, Report A57, 2000, Available at: www.cis.hut.fi/projects/somtoolbox/package/papers/techrep.pdf
}
}
\seealso{
db_indexR
}
\author{
Sabina Licen, Pierluigi Barbieri
}
